backend/requirements.txt
Django
psycopg2-binary
djangorestframework
Pillow
stripe
django-redis
celery